feat: add cooldown feature to paths, calculate fees based on msgs
We use a cooldown feature on paths to prevent the bot for overtrading a
path whiles the arb is already gone. if it did one attempt cooldown is
set to true, on next block arrival the bot sets all paths cooldowns back
to false.
Also, this commit changes the way fees are attached to TXs, because fees
do not only scale with amount of pools, but also amount of messages used
for a pooltrade, therefore the amount of msgs is used as an indexer for
the amount of txfees to attach to a tx