Release (#63)
* feat: protocol fee handling, outgivenin/stateupdate calc change

BREAKING: this commit requires an extra field on the pool environment
variables: LPratio

* feat: add cooldown feature to paths, calculate fees based on msgs

We use a cooldown feature on paths to prevent the bot for overtrading a
path whiles the arb is already gone. if it did one attempt cooldown is
set to true, on next block arrival the bot sets all paths cooldowns back
to false.
Also, this commit changes the way fees are attached to TXs, because fees
do not only scale with amount of pools, but also amount of messages used
for a pooltrade, therefore the amount of msgs is used as an indexer for
the amount of txfees to attach to a tx

* bug: fix profit calculation based on tradesize and skipbid

* bug: added consoles to check applyTradeOnPool update

* fix: removed extra file slacklogger.ts

* chore: Added clean step to package.json

* feat: Added docker-compose.yml

* fix: Updated dockerfile for optimization and docker-compose compatibility

* chore: Updated dev dependencies

* fix: Updated allow for .env file to be read from docker-compose

* Sign of Life in Minutes

Sign of Life in Minutes to be more consistent when running on Juno or Injective

* var names

* Output String

* chore: update packages align with release

* chore: delete comments and newlines

* chore: extra cooldown=true removed

* bug: fix send message processing

* chore: use isSwapMessage in mempool and isDefaultSwapMessage fn

* Add Telegram Logger and remove "clear"

* Remove unnecessary console.log

Remove unnecessary console.log from testing

* eslint telegram logger

* Add externalExemptCodes to env & external config

* Add EXTERNAL_EXEMPT_CODES to example .env

* chore: update examples and add skip settings to modules

* cicd: enable linting in package.json's scripts

* chore: change flashloan fee to 0.15% in examples

* chore: add eslinter to prebuild script

* chore: change flashloan fee in terra example

* Update Dockerfile: add eslint copy

* Update package.json

---------

Co-authored-by: CarlTheLooper <carlthelooper@gmail.com>
Co-authored-by: nick134 <nick134-bit@proton.me>
Co-authored-by: zJ <zzju@protonmail.com>